+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## Changelog
-- Reworked the Network part, now it can run directly without relying on the Forgified Fabric API.
+- Update Forgified Fabric API version.
 - update NeoForge `26.1.2.92` -> `26.1.2.103`
 - sync upstream